NetBSD-macppc ctype needs uchars.

+++ src/cmd/9660/dump.c
@@ -456,7 +456,7 @@ static int
 isalldigit(char *s)
 {
 	while(*s)
-		if(!isdigit(*s++))
+		if(!isdigit((uchar)*s++))
 			return 0;
 	return 1;
 }
